third_party/WebKit/LayoutTests/scrollbars/iframe-scrollbar-becomes-custom-expected.txt

PASS 250 is 250
PASS width < origWidth is true
PASS successfullyParsed is true

TEST COMPLETE
